[config] Reorder config crate and move test to tests.rs
